Add files via upload
This is a Catch and Score game made in assembly language.
The game works with shift left and shift right keys (for left and right movement of bucket respectively). Each block caught will result in incrementing score and catching the bomb will result in game over. Moreover, the game has a time limit of two minutes.